Move `mod/` to `mods/`
