+++ src/models/book.js
@@ -1,5 +1,5 @@
 import { Model } from 'backbone'
-import _ from 'mod/utils'
+import _ from 'mods/utils'
 
 export default Model.extend({
